mcp_servers/thanos_mcp/requirements.txt
fastmcp>=0.1.0
httpx>=0.23.0
uvicorn>=0.18.0
python-dotenv>=0.19.0
PyYAML>=6.0
prometheus-api-client>=0.5.0